subtituição do hashMap por List

+++ Macaco.java
@@ -3,9 +3,9 @@
 public class Macaco {
     
     private int id, envP, envI;
-    private long nPar, nImp;
+    private int nPar, nImp;
    
-    public Macaco(int id, int envP, int envI, long nPar, long nImp){
+    public Macaco(int id, int envP, int envI, int nPar, int nImp){
         this.id = id;
         this.envP = envP;
         this.envI = envI;       
@@ -16,13 +16,13 @@ public class Macaco {
     public int getId(){ return id; }
     public int getEnvP(){ return envP; }
     public int getEnvI(){ return envI; }
-    public long getPar(){ return nPar; }
-    public long getImp(){ return nImp; }
-    public long getTam(){ return nPar+nImp; }
+    public int getPar(){ return nPar; }
+    public int getImp(){ return nImp; }
+    public int getTam(){ return nPar+nImp; }
     public void clear(){ nPar = 0; nImp = 0; }
 
-    public void addPar(long p){  nPar+= p; }
-    public void addImp(long i){  nImp+= i; }
+    public void addPar(int p){  nPar+= p; }
+    public void addImp(int i){  nImp+= i; }
    
 
     @Override
